% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile_css.R
\name{compile_css}
\alias{compile_css}
\title{Compile a scss file to a css output}
\usage{
compile_css(scss_file, output_path, output_style = "compressed")
}
\arguments{
\item{scss_file}{Character string of the scss file to use for compilation.}

\item{output_path}{Specifies the location of the output file.}

\item{output_style}{Bracketing and formatting style of the CSS output. Possible styles: "nested", "expanded", "compact", and "compressed".}
}
\value{

}
\description{
Compile a scss file to a css output
}
\examples{
\dontrun{
compile_css("scss/html_document.scss", "inst/resources/css/bootstrap.css", "compressed")
}
}
