% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unhcr_bslib.R
\docType{data}
\name{unhcr_bslib}
\alias{unhcr_bslib}
\title{UNHCR bslib variables}
\format{
A list
}
\usage{
unhcr_bslib
}
\description{
Bootstrap default to be used in bslib for UNHCR
}
\examples{
unhcr_bslib
}
\keyword{datasets}
