% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unhcr_css_files.R
\name{unhcr_css_files}
\alias{unhcr_css_files}
\title{List available CSS files}
\usage{
unhcr_css_files()
}
\description{
Print a list of available CSS files included in unhcrdesign
}
\examples{
unhcr_css_files()
}
