% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_unhcr_css.R
\name{use_unhcr_css}
\alias{use_unhcr_css}
\title{CSS file location}
\usage{
use_unhcr_css(css_file)
}
\arguments{
\item{css_file}{Name of css file to retrieve}
}
\description{
Returns the location of css file(s)
}
