% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_unhcr_logo.R
\name{use_unhcr_logo}
\alias{use_unhcr_logo}
\title{Use UNHCR Logo}
\usage{
use_unhcr_logo(logo = c("blue", "white", "black"), data_uri = FALSE)
}
\arguments{
\item{logo}{Logo type}

\item{data_uri}{Logical.  Should a data uri be returned?
If FALSE, file location is returned.}
}
\description{
Returns location or data uri of a UNHCR logo
}
