% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_unhcr_scss.R
\name{use_unhcr_scss}
\alias{use_unhcr_scss}
\title{SCSS file location}
\usage{
use_unhcr_scss(scss_file)
}
\arguments{
\item{scss_file}{Name of scss file to retrieve}
}
\description{
Returns the location of scss file(s)
}
